Remove feedback generator helper
