+++ Sources/Charcoal/Filters/Range/Slider/RangeSliderView.swift
@@ -229,7 +229,8 @@ extension RangeSliderView: StepSliderDelegate {
 
     func stepSlider(_ stepSlider: StepSlider, didChangeStep step: Step) {
         if lowValueSlider.step == highValueSlider.step, generatesHapticFeedbackOnValueChange {
-            FeedbackGenerator.generate(.collision)
+            let generator = UIImpactFeedbackGenerator(style: .heavy)
+            generator.impactOccurred()
         }
 
         handleStepUpdate(newStep: step, stepSlider: stepSlider)
